Boundaries and initial player placement updated
Updated the boundary values to be more accurate and changed the initial spawn in coordinates of the player to be in the middle of the map as we currently have planned. These values may need adjusting when the proper player assets are brought in and the game is developed further.