Centred main menu for current resolution + Updated window text
Changed the locations of the main menu text and buttons and controls screen text to be more centred for the current resolution we have set for the game window. The game window text has also been updated to reflect the planned name we have for the game instead of "base game".