Setting up player animations + Fixing Health Bar
Setting up the player animations so they correspond with the key the player is pressing, there is set up for attacking animations when combat is added into the game. Please note that spamming keys will throw an error, this is something to look into when we can.
The spritesheet class can also be used to read enemy spritesheets when we get to setting those up. On top of this, updated the health bar to fit in the bar and float more centrally above the player's head.