Enemies following the player
Main update: The enemies can now follow the player from any direction and seem to be able to locate the player from different places on the map.  When thinking of adding groups/waves, multiple instances of the enemy will need to be created/called (which can be done through loops hopefully) and we can also change the speed of each enemy so we can maybe implement a random value for a bit of variation. One thing that needs adding is enemy health which will come with the combat update.

On top of this, I made the "Player" its own class. This may need updating or changing as the development of the game continues.