Big Update - Inventory / Shop / Chat UI
Quite a big update this one, I may miss out some of the work that's been done. Since these features are quite interlinked, it's been tricky trying to split the update up into different ones so I think a "Big Update" accounts for this.

The inventory system has been updated to now be able to have items added and removed from it, these items being "Health Potion", "Strength Potion" and "Key"; the health potion and strength potion already appear within the inventory from the start of the game with a "0" counter next to them and the key item will be appended to the inventory when it is bought. With every purchase from the shop, the matching item's count increases and then these items can be used within the inventory (apart from the key due to it being a story progressing item) which then does it's intended purpose and decreases the item count. [The inventory can still be accessed with 'E' at this point in development]

A coin counter now sits at the top right of the screen and does need altering a bit as it's still default text but it's functionality seems to work, updating whenever the player purchases an item from the shop. I'm unsure of it's functionality when it comes to adding coins as this has not been worked on yet and will be worked on when enemies are a bit more finalised since they will be the ones to be dropping coins or just adding a value to the coin count.

The shop system and chat UI for the shop has been added now, with a simple layout ready for content (writing and images) and buttons that allow for the player to purchase items if they have enough coins. Although this hasn't been added in yet, it's planned that defeating enemies will either grant / drop coins so the player can increase their coin count. In order to access the shop, the player must be within a certain radius surrounding the shop sprite and press the "i" key (this is subject to change as development continues) and they can close the shop with the same key as long as they are in range. This might need taking a look at because if they aren't in range and are attempting to close the shop menu it might become annoying (maybe an automatic open/close instead of with a key?). There needs to be more development with this chat UI as story elements / "chatting" and the shopkeeper needs to be added in. It's key to note that in order to be able to implement the wave system, a counter or boolean needs to be set up to let the game know when the player has talked to the shopkeeper at least once so then it knows to start a timer until the enemies start spawning.