Player Score + Start of Updating Screens
Added in a scoring system for the player which will be shown to them on their death or completion of the game. If they beat the game, this value is multiplied based upon their remaining health as a "bonus". On top of this, the screens have started to be updated into their more finalised state with backgrounds. The buttons and text still need updating though.