Cutscenes / Dialogue + Shop Keeper
Added 4 "cutscenes" or dialogue sections into the game to add more story elements. During these "cutscenes", the player will not be able to move (the controls cannot be used). These cutscenes happen at: the start of the game, when the player first interacts with the shop, when the second level is started and after the boss is defeated.

On top of this, the sprite/animations for the shopkeeper have been added and implemented into the appropriate areas in the game.