Starting to Add Music + Sounds (important note)
Starting to add music and sounds to the game. However, replit doesn't support the use of the audio mixer so this code cannot be tested unless working in an environment like vscode probably. I'm pushing this so the code is available (its just commented out), and this will likely be the last thing that gets worked on so we can still gather game evidence from this source if needed.